cambios para admitir los dtos en admin y alumno (no se puede en angular)

+++ src/main/java/com/fleming/bancodelibros/repos/AlumnoRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 import com.fleming.bancodelibros.modelo.Alumno;
 
 @Repository
-public interface AlumnoRepository extends JpaRepository<Alumno, Integer>{
+public interface AlumnoRepository extends JpaRepository<Alumno, Long>{
 }
